backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.13
pinecone-client==3.0.0
sentence-transformers==2.2.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
lxml==4.9.3
aiofiles==23.2.1